+++ pdf2md/server/Cargo.toml
@@ -44,6 +44,7 @@ openai_dive = "0.6.6"
 env_logger = "0.11.5"
 utoipa-actix-web = "0.1.2"
 futures = "0.3.31"
+regex = "1.11.1"
 
 [features]
 default = []
