+++ pdf2md/server/ch_migrations/1731447246_remove_chunks_processed_field/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE file_tasks ADD COLUMN IF NOT EXISTS chunks_processed UInt32;
