+++ pdf2md/server/src/errors.rs
