+++ pdf2md/server/src/routes/get_task.rs
@@ -7,7 +7,7 @@ use crate::{
 
 /// Retieve a File Task by ID
 ///
-/// This endpoint retrieves a task by its id. The task is returned along with the chunks that have been created, if the file chunking has been completed.
+/// This endpoint retrieves a task by its id. The task is returned along with the pages that have been created, if the file chunking has been completed.
 #[utoipa::path(
     get,
     path = "/task/{task_id}",
@@ -15,11 +15,11 @@ use crate::{
     context_path = "/api",
      params(
         ("task_id" = uuid::Uuid, Path, description = "The id of the task you want to retrieve."),
-        ("limit" = Option<u32>, Query, description = "The number of chunks to return."),
+        ("limit" = Option<u32>, Query, description = "The number of pages to return."),
         ("pagination_token" = Option<String>, Query, description = "The pagination token to use for the next request."),
     ),
     responses(
-        (status = 200, description = "JSON response payload containing the created chunks", body = models::GetTaskResponse),
+        (status = 200, description = "JSON response payload containing the created pages", body = models::GetTaskResponse),
         (status = 400, description = "Error typically due to deserialization issues", body = ErrorResponseBody),
     ),
     security(
@@ -34,7 +34,7 @@ async fn get_task(
 ) -> Result<HttpResponse, ServiceError> {
     let task_id = task_id.into_inner();
     let task = crate::operators::clickhouse::get_task(task_id, &clickhouse_client).await?;
-    let result = crate::operators::clickhouse::get_task_chunks(
+    let result = crate::operators::clickhouse::get_task_pages(
         task,
         data.limit,
         data.pagination_token,
